Promote 36 Goomboss shards into src/actors/Goomboss.cpp (ov074, partial)
Folds ROM ordinals 2..37 of ov074's daKuriKing_c run into one translation
unit, .text 0x0211f154..0x02121380, 8748 bytes.  36/36 MATCH, objisolate
clean, reloc-destinations clean; linkcheck SCRATCH-LINK-VERIFIED with
106/106 modules exact and the full ROM byte-identical to stock.

The claim is a subset of the class's 55-function run, and the cut is forced
twice over.

  * func_ov074_02121380 is a banked near-miss (11 divergent words,
    notes/mwccarm-codegen.md 6bq) with no delinks.txt entry, so the
    cartridge's own bytes still cover 0x02121380..0x021216f4 and the run has
    a hole in it.  No delink block and no TU manifest in this tree carries
    two .text runs, so only one contiguous side is claimable; this is the
    larger, 38 functions against 16.

  * The destructor pair is excluded for the key function, not for the
    destructor order.  ROM order is D1 (0x0211f000) then D0 (0x0211f0a0),
    which is the reproducible direction and needs an inline in-class
    destructor -- but inline, the key function moves to the first
    out-of-line virtual the class declares, Behavior at 0x02121bf0, on the
    far side of the hole.  A TU that owns no key function emits no vtable,
    nothing then odr-uses the destructor, and neither variant is emitted.  A
    forcing scaffold does emit them in ROM order when `delete p;` precedes
    `p->~Goomboss();`, but mwccarm lays the scaffold into .text immediately
    below D1, and D1 is the first byte of ov074's .text.  There is no
    address below it.

Match func_ov074_02121380 and the hole closes: one TU would own Behavior and
therefore the key function, the vtable and both destructors, and the whole
0x0211f000..0x0212290c run would promote.

Two reconciliations the merge forced.  common.h is now included before
Goomboss.h, because Goomboss.h reaches math/Matrix.h's structured Matrix4x3
while every legacy shard saw common.h's flat s32 m[12].  And
func_ov074_021201f0's angle temp is spelt `struct Ang16 { short v[3]; }`:
its shard was a .c file where the whole-struct assignment is a block move,
while C++ scalarises it and the body came out 0x20 short.

compiler_only_output is one row, Vector3's inline destructor.  The vtable
and the five-level RTTI chain are not emitted here at all, because the
licensed range excludes the key function.  externalized_output is absent.

ov074 had no sibling oracle, so notes/data/class-facts/Goomboss.json records
the measured chain -- daKuriKing_c -> dEnemyBase_c -> dActor_c -> dBase_c ->
fBase_c, five levels -- and the 11 data rows a key-function-owning TU emits,
which is the row-count formula's prediction exactly.  It also records the
name trap: the cartridge's RTTI spells this class daKuriKing_c while the
vtable symbol is _ZTV8Goomboss, so rtti_vtables.py answers only to the ROM
name and romdata_check reports the coined _ZTI/_ZTS as UNNAMED.
