+++ config/arm9/overlays/ov074/delinks.txt
@@ -12,149 +12,9 @@ src/_ZN8GoombossD0Ev.cpp:
     complete
     .text start:0x0211f0a0 end:0x0211f154
 
-src/func_ov074_0211f154.c:
+src/actors/Goomboss.cpp:
     complete
-    .text start:0x0211f154 end:0x0211f244
-
-src/func_ov074_0211f244.cpp:
-    complete
-    .text start:0x0211f244 end:0x0211f344
-
-src/func_ov074_0211f344.c:
-    complete
-    .text start:0x0211f344 end:0x0211f38c
-
-src/func_ov074_0211f38c.c:
-    complete
-    .text start:0x0211f38c end:0x0211f5b8
-
-src/func_ov074_0211f5b8.cpp:
-    complete
-    .text start:0x0211f5b8 end:0x0211fa08
-
-src/func_ov074_0211fa08.c:
-    complete
-    .text start:0x0211fa08 end:0x0211fa74
-
-src/func_ov074_0211fa74.c:
-    complete
-    .text start:0x0211fa74 end:0x0211fb84
-
-src/func_ov074_0211fb84.c:
-    complete
-    .text start:0x0211fb84 end:0x0211fbd0
-
-src/func_ov074_0211fbd0.c:
-    complete
-    .text start:0x0211fbd0 end:0x0211fc34
-
-src/func_ov074_0211fc34.c:
-    complete
-    .text start:0x0211fc34 end:0x0211fc38
-
-src/func_ov074_0211fc38.c:
-    complete
-    .text start:0x0211fc38 end:0x0211fd48
-
-src/func_ov074_0211fd48.c:
-    complete
-    .text start:0x0211fd48 end:0x0211fd74
-
-src/func_ov074_0211fd74.cpp:
-    complete
-    .text start:0x0211fd74 end:0x0211ffac
-
-src/func_ov074_0211ffac.c:
-    complete
-    .text start:0x0211ffac end:0x0211ffcc
-
-src/func_ov074_0211ffcc.c:
-    complete
-    .text start:0x0211ffcc end:0x0212007c
-
-src/func_ov074_0212007c.c:
-    complete
-    .text start:0x0212007c end:0x02120080
-
-src/func_ov074_02120080.cpp:
-    complete
-    .text start:0x02120080 end:0x0212016c
-
-src/func_ov074_0212016c.c:
-    complete
-    .text start:0x0212016c end:0x0212018c
-
-src/func_ov074_0212018c.cpp:
-    complete
-    .text start:0x0212018c end:0x021201ec
-
-src/func_ov074_021201ec.c:
-    complete
-    .text start:0x021201ec end:0x021201f0
-
-src/func_ov074_021201f0.c:
-    complete
-    .text start:0x021201f0 end:0x021203e0
-
-src/func_ov074_021203e0.c:
-    complete
-    .text start:0x021203e0 end:0x021203e4
-
-src/func_ov074_021203e4.cpp:
-    complete
-    .text start:0x021203e4 end:0x0212042c
-
-src/func_ov074_0212042c.cpp:
-    complete
-    .text start:0x0212042c end:0x02120474
-
-src/func_ov074_02120474.cpp:
-    complete
-    .text start:0x02120474 end:0x021204c0
-
-src/func_ov074_021204c0.c:
-    complete
-    .text start:0x021204c0 end:0x021206c8
-
-src/func_ov074_021206c8.c:
-    complete
-    .text start:0x021206c8 end:0x021207b8
-
-src/func_ov074_021207b8.c:
-    complete
-    .text start:0x021207b8 end:0x02120808
-
-src/func_ov074_02120808.cpp:
-    complete
-    .text start:0x02120808 end:0x0212087c
-
-src/func_ov074_0212087c.c:
-    complete
-    .text start:0x0212087c end:0x02120b24
-
-src/func_ov074_02120b24.c:
-    complete
-    .text start:0x02120b24 end:0x02120b90
-
-src/func_ov074_02120b90.c:
-    complete
-    .text start:0x02120b90 end:0x02120bb8
-
-src/func_ov074_02120bb8.c:
-    complete
-    .text start:0x02120bb8 end:0x02120d74
-
-src/func_ov074_02120d74.c:
-    complete
-    .text start:0x02120d74 end:0x02121270
-
-src/func_ov074_02121270.c:
-    complete
-    .text start:0x02121270 end:0x02121300
-
-src/func_ov074_02121300.cpp:
-    complete
-    .text start:0x02121300 end:0x02121380
+    .text start:0x0211f154 end:0x02121380
 
 src/func_ov074_021216f4.c:
     complete
