+++ config/match_provenance.jsonl
@@ -16,7 +16,7 @@
 {"id":"ov071:0x0211f7d4","module":"ov071","addr":34731988,"name":"func_ov071_0211f7d4","matchProvenance":{"kind":"ai","model":"grok-4.5","reasoning":"high","harness":"grok-build"},"author":"lunavyqo","srcPath":"src/func_ov071_0211f7d4.c"}
 {"id":"ov006:0x020e2dbc","module":"ov006","addr":34483644,"name":"func_ov006_020e2dbc","matchProvenance":{"kind":"ai","model":"grok-4.5","reasoning":"high","harness":"grok-build"},"author":"lunavyqo","srcPath":"src/func_ov006_020e2dbc.c"}
 {"id":"ov007:0x020c9110","module":"ov007","addr":34378000,"name":"func_ov007_020c9110","matchProvenance":{"kind":"ai","model":"grok-4.5","reasoning":"high","harness":"grok-build"},"author":"lunavyqo","srcPath":"src/func_ov007_020c9110.c"}
-{"id":"ov074:0x0211f244","module":"ov074","addr":34730564,"name":"func_ov074_0211f244","matchProvenance":{"kind":"ai","model":"grok-4.5","reasoning":"high","harness":"grok-build"},"author":"lunavyqo","srcPath":"src/func_ov074_0211f244.c"}
+{"id":"ov074:0x0211f244","module":"ov074","addr":34730564,"name":"func_ov074_0211f244","matchProvenance":{"kind":"ai","model":"grok-4.5","reasoning":"high","harness":"grok-build"},"author":"lunavyqo","srcPath":"src/actors/Goomboss.cpp"}
 {"id":"ov006:0x020e2f78","module":"ov006","addr":34484088,"name":"func_ov006_020e2f78","matchProvenance":{"kind":"ai","model":"grok-4.5","reasoning":"high","harness":"grok-build"},"author":"lunavyqo","srcPath":"src/func_ov006_020e2f78.c"}
 {"id":"ov007:0x020c25fc","module":"ov007","addr":34350588,"name":"func_ov007_020c25fc","matchProvenance":{"kind":"ai","model":"grok-4.5","reasoning":"high","harness":"grok-build"},"author":"lunavyqo","srcPath":"src/func_ov007_020c25fc.c"}
 {"id":"ov002:0x020f39ec","module":"ov002","addr":34552300,"name":"func_ov002_020f39ec","matchProvenance":{"kind":"ai","model":"grok-4.5","reasoning":"high","harness":"grok-build"},"author":"lunavyqo","srcPath":"src/func_ov002_020f39ec.c"}
@@ -687,8 +687,8 @@
 {"id":"ov075:0x0211a948","module":"ov075","addr":34711880,"name":"func_ov075_0211a948","matchProvenance":{"kind":"ai","model":"claude-opus-5","reasoning":"high","harness":"claude-code"},"author":"tangosdev","srcPath":"src/func_ov075_0211a948.c"}
 {"id":"ov075:0x02119dc4","module":"ov075","addr":34708932,"name":"func_ov075_02119dc4","matchProvenance":{"kind":"ai","model":"claude-opus-5","reasoning":"high","harness":"claude-code"},"author":"tangosdev","srcPath":"src/func_ov075_02119dc4.cpp"}
 {"id":"ov075:0x02117d80","module":"ov075","addr":34700672,"name":"func_ov075_02117d80","matchProvenance":{"kind":"ai","model":"claude-opus-5","reasoning":"high","harness":"claude-code"},"author":"andrewboudreau","srcPath":"src/func_ov075_02117d80.cpp"}
-{"id":"ov074:0x021201f0","module":"ov074","addr":34734576,"name":"func_ov074_021201f0","matchProvenance":{"kind":"ai","model":"claude-opus-5","reasoning":"high","harness":"claude-code"},"author":"tangosdev","srcPath":"src/func_ov074_021201f0.c"}
-{"id":"ov074:0x021204c0","module":"ov074","addr":34735296,"name":"func_ov074_021204c0","matchProvenance":{"kind":"ai","model":"claude-opus-5","reasoning":"high","harness":"claude-code"},"author":"tangosdev","srcPath":"src/func_ov074_021204c0.c"}
+{"id":"ov074:0x021201f0","module":"ov074","addr":34734576,"name":"func_ov074_021201f0","matchProvenance":{"kind":"ai","model":"claude-opus-5","reasoning":"high","harness":"claude-code"},"author":"tangosdev","srcPath":"src/actors/Goomboss.cpp"}
+{"id":"ov074:0x021204c0","module":"ov074","addr":34735296,"name":"func_ov074_021204c0","matchProvenance":{"kind":"ai","model":"claude-opus-5","reasoning":"high","harness":"claude-code"},"author":"tangosdev","srcPath":"src/actors/Goomboss.cpp"}
 {"id":"ov100:0x021458d4","module":"ov100","addr":34887892,"name":"_ZN12daStarGate_cD1Ev","matchProvenance":{"kind":"ai","model":"gpt-5","reasoning":"high","harness":"codex"},"author":"tangosdev","srcPath":"src/_ZN12daStarGate_cD1Ev.cpp"}
 {"id":"ov100:0x02145904","module":"ov100","addr":34887940,"name":"_ZN12daStarGate_cD0Ev","matchProvenance":{"kind":"ai","model":"gpt-5","reasoning":"high","harness":"codex"},"author":"tangosdev","srcPath":"src/_ZN12daStarGate_cD0Ev.cpp"}
 {"id":"ov100:0x02145948","module":"ov100","addr":34888008,"name":"_ZN12daStarGate_c18St_StayClosed_MainEP6Player","matchProvenance":{"kind":"ai","model":"gpt-5","reasoning":"high","harness":"codex"},"author":"tangosdev","srcPath":"src/_ZN12daStarGate_c18St_StayClosed_MainEP6Player.cpp"}
