+++ src/actors/Goomboss.cpp
@@ -1,0 +1,1499 @@
+//cpp
+/* ov074/Goomboss -- 36 of the class's 55 shards, folded into the translation
+ * unit Nintendo shipped.  The ROM RTTI name of this class is daKuriKing_c
+ * (_ZTI12daKuriKing_c at ov074:0x02122e30); `Goomboss` is the tree's coined
+ * spelling and every symbol in config/ uses it, so the file keeps it.
+ *
+ * THIS IS A PARTIAL, AND THE CUT IS NOT NEGOTIABLE.  The class's linker run is
+ * 0x0211f000..0x0212290c, but func_ov074_02121380 (0x02121380, size 0x374) is a
+ * banked near-miss: it has no delinks.txt entry anywhere, so the cartridge's own
+ * bytes cover 0x02121380..0x021216f4 and a licensed claim cannot have a hole.
+ * The run therefore splits into a 38-function lower side and a 16-function upper
+ * side, and this file takes the larger one -- less its first two members:
+ *
+ *   ordinals 0-1  _ZN8GoombossD1Ev / _ZN8GoombossD0Ev.  The cartridge orders
+ *     them D1 (0x0211f000) then D0 (0x0211f0a0), which mwccarm 2004/b56 emits
+ *     only from an inline in-class destructor -- an out-of-line definition emits
+ *     D2, D0, D1 (measured).  But with the destructor inline the class's key
+ *     function moves to the first out-of-line virtual it declares, Behavior at
+ *     0x02121bf0, which is on the FAR side of the hole.  This TU then owns no
+ *     key function, emits no vtable, and nothing odr-uses the destructor, so
+ *     neither variant is emitted at all (measured: an empty TU over the inline
+ *     header emits zero symbols).  A forcing scaffold does emit the pair in ROM
+ *     order -- `delete p;` before `p->~Goomboss();`, because the variants come
+ *     out in reverse order of first use -- but the scaffold is itself emitted
+ *     into .text immediately BELOW D1, and D1 sits at 0x0211f000, the first byte
+ *     of ov074's .text.  There is no address below it.  So the pair cannot live
+ *     in this TU's licensed run; src/_ZN8GoombossD1Ev.cpp and
+ *     src/_ZN8GoombossD0Ev.cpp keep their own delinks entries.
+ *
+ *   ordinal 38 and above stay in their own shards for the reason above.
+ *
+ * Licensed run: .text 0x0211f154..0x02121380, 36 functions, 36/36 MATCH.
+ *
+ * FUNCTION ORDER IS DELIBERATELY THE REVERSE OF THE ROM'S -- mwccarm 2004/b56
+ * emits one .text section per function in the REVERSE of source order, so the
+ * highest-address ROM function is written first.  Do not reorder.
+ *
+ * Two reconciliations the merge forced, both documented mwccarm behaviours:
+ *   - common.h is included BEFORE Goomboss.h.  Goomboss.h reaches
+ *     math/Matrix.h's structured Matrix4x3 (Matrix3x3 r; Vector3 t) while every
+ *     legacy shard saw common.h's flat `s32 m[12]`.  With the structured
+ *     spelling first, func_ov074_02121270 and func_ov074_02121300 both went long
+ *     on their whole-matrix assignments; the reorder restored 2/2.
+ *   - func_ov074_021201f0's angle temp is spelt `struct Ang16 { short v[3]; }`.
+ *     Its shard is a .c file compiled as C, where the whole-struct assignment is
+ *     a block move; as C++ the same assignment is scalarised, the local loses
+ *     its frame home and the body came out 0x20 short.  The array member
+ *     restores the block move.
+ *
+ * Assembled from src/func_ov074_0211f154.c .. src/func_ov074_02121300.cpp,
+ * ROM ordinals 2..37 of the class's linker run.
+ */
+
+/* Includes: union of the legacy files', first-seen in ROM-ascending
+ * processing order. NOT verified for header ordering constraints (e.g. a
+ * common.h-before-X rule) -- watch for new compile errors after this. */
+#include "common.h"   /* BEFORE Goomboss.h: see the note above */
+#include "Goomboss.h"
+#include "types.h"
+#include "decl_common.h"
+#include "SharedFilePtr.h"
+#include "MaterialChanger.h"
+#include "TextureSequence.h"
+#include "dBgCh_Gnd.h"
+
+/* Local shadow declarations carried from the legacy files verbatim.
+ * NOT reconciled against real project headers -- check include/*.h for
+ * each of these before compiling; a real header should usually win. */
+/* shadow struct 'Anim' */
+struct Anim {
+    char _pad0[4];
+};
+
+/* shadow struct 'Obj' */
+struct Obj {
+    char _pad0[0x8e];
+    s16 angle;            /* 0x8e */
+    char _pad1[0x260 - 0x90];
+    struct Anim anim;     /* 0x260 */
+};
+
+/* shadow typedef 'Inner' */
+typedef struct { int field0; int field4; } Inner;
+
+/* shadow typedef 'Entry' */
+typedef struct { Inner *ptr; int a; int b; } Entry;
+
+/* shadow struct 'Sub' */
+struct Sub {
+  virtual void v0(); virtual void v1(); virtual void v2();
+  virtual void v3(); virtual void v4(); virtual void m(void*);
+};
+
+/* shadow struct 'BMD_File' */
+struct BMD_File;
+
+/* shadow struct 'BCA_File' */
+struct BCA_File;
+
+/* shadow struct 'BTP_File' */
+struct BTP_File;
+
+/* shadow struct 'BMA_File' */
+struct BMA_File;
+
+/* shadow struct 'BTA_File' */
+struct BTA_File;
+
+/* shadow struct 'Vector3' */
+struct Vector3;
+
+/* shadow struct 'Vector3_16' */
+struct Vector3_16;
+
+/* shadow struct 'Obj210' */
+struct Obj210 {
+    virtual void v0();
+    virtual void v1();
+    virtual void v2();
+    virtual void v3();
+    virtual void v4();
+    virtual void v5(int a);
+};
+
+
+
+#define AT(p,off) ((void*)(int)((char*)(p)+(off)))
+#define LI(v) ((int)((long long)(v)))
+#define false 0
+#define FX(a, b) (int)(((long long)(a) * (long long)(b) + 0x800) >> 12)
+
+extern "C" {
+extern void _ZN6Camera9SetFlag_3Ev(void* cam);
+extern void Matrix4x3_FromRotationY(void* m, short ang);
+extern void MulVec3Mat4x3(void* a, void* m, void* b);
+extern void _ZN6Camera9SetLookAtERK7Vector3(void* cam, struct Vector3* v);
+extern void _ZN6Camera6SetPosERK7Vector3(void* cam, struct Vector3* v);
+extern char* data_0209f318;
+extern "C" int RandomIntInternal(int *seed);
+extern "C" int data_0209e650;
+extern void* _ZN8dActor_c15FindWithActorIDEjPS_(unsigned int, void*);
+extern u16 DecIfAbove0_Short(u16* p);
+extern s16 Vec3_VertAngle(const struct Vector3* v1, const struct Vector3* v0);
+extern char* _ZN8dActor_c5SpawnEjjRK7Vector3PK10Vector3_16as(u32 a, u32 b, const struct Vector3* pos, const struct Vector3_16* ang, int e, int f);
+extern char* _ZN8dActor_c13ClosestPlayerEv(void* self);
+extern s16 data_02082214[];
+extern int _ZN9Animation8FinishedEv(void* c);
+extern int _ZN6Player12GetTalkStateEv(void* p);
+extern int _ZN6Player11ShowMessageER7fBase_cjPK7Vector3hh(void* p, void* actor, u32 id, const Vector3* v, u32 a, u32 b);
+extern void _ZN8dActor_c10PoofDustAtERK7Vector3(void* c, const Vector3* v);
+extern void _ZN7fBase_c18MarkForDestructionEv(void* a);
+extern void func_02012694(u32 a, void* b);
+extern int _ZN6Player18HasFinishedTalkingEv(void* p);
+extern void _ZN8Particle6System9NewSimpleEj5Fix12IiES2_S2_(u32 id, int x, int y, int z);
+extern void _ZN7Message7EndTalkEv();
+extern void _ZN5Sound22StopLoadedMusic_Layer3Ev();
+extern void func_02011cfc();
+extern void _ZN5Sound17ChangeMusicVolumeEj5Fix12IiE(u32 a, int f);
+extern int func_01ffa344(int);
+extern unsigned short data_ov074_02122e04[];
+extern int ApproachAngle(void*,int,int,int,int);
+extern int _ZN6Player9StartTalkER7fBase_cb(void* p, void* a, int b);
+extern void _ZN7Message11PrepareTalkEv(void);
+void _ZN8dActor_c19DisappearPoofDustAtERK7Vector3(void* self, void* v);
+void func_ov084_021296b0(void* a, void* b);
+int __aeabi_idiv(int a, int b);
+extern int _ZN8dActor_c13DistToCPlayerEv(char* c);
+extern int _ZN8dActor_c18HorzAngleToCPlayerEv(char* c);
+extern int AngleDiff(int a, int b);
+extern void _ZN5Sound22LoadAndSetMusic_Layer3Ej(unsigned int id);
+extern int func_01ffa594(int a, int b);
+extern int func_01ffa4bc(int a);
+extern int func_01ff9378(int a, int b);
+extern int _ZN4cstd4fdivEii(int a, int b);
+extern int Math_Function_0203b14c(int* p, int a, int b, int d, int e);
+extern unsigned short data_ov074_02122dfc[];
+extern int data_ov074_0212291c[];
+extern void Vec3_Sub(Vector3 *out, const Vector3 *a, const Vector3 *b);
+extern s32 Vec3_HorzLen(const Vector3 *v);
+extern s16 _ZN4cstd5atan2E5Fix12IiES1_(s32 y, s32 x);
+extern unsigned char DecIfAbove0_Byte(unsigned char *p);
+extern char *_ZN8dActor_c10FindWithIDEj(unsigned int id);
+extern void _ZN6Player12Unk_020c6a10Ej(char *p, unsigned int a);
+extern void _ZN6Player4HurtERK7Vector3j5Fix12IiEjjj(char *p, const struct Vector3 *v, unsigned int a, int b, unsigned int c, unsigned int d, unsigned int e);
+extern s16 Vec3_HorzAngle(const struct Vector3 *v0, const struct Vector3 *v1);
+extern void MulMat4x3Mat4x3(void* out, void* a, void* b);
+extern void Vec3_LslInPlace(void* v, int n);
+extern struct Matrix4x3 data_020a0e68;
+extern int _ZNK10dBgCh_Actr10IsOnGroundEv(void *c);
+extern struct Matrix4x3 IDENTITY_MATRIX4X3;
+extern void _ZN9Animation7AdvanceEv(void* p);
+void _ZN9ModelAnim7SetAnimEP8BCA_Filei5Fix12IiEj(void* anim, void* file, int a, int b, unsigned int u);
+void _ZN18TextureTransformer7SetFileER8BTA_Filei5Fix12IiEj(void* tt, void* file, int a, int b, unsigned int u);
+extern char data_ov074_02122f3c[];
+extern char data_ov074_02122f38[];
+void UnloadKeyModels(int i);
+extern char data_ov002_0210da30;
+extern char data_ov084_02130cf8;
+extern char data_ov074_02123000;
+extern char data_ov074_02123040;
+extern void _ZN15TextureSequence6UpdateER15ModelComponents(void* a, void* b);
+extern void _ZN15MaterialChanger6UpdateER15ModelComponents(void* a, void* b);
+extern void _ZN18TextureTransformer6UpdateER15ModelComponents(void* a, void* b);
+extern void _ZN8dActor_c9UpdatePosEP5dCc_c(void *thiz, void *clsn);
+extern void _ZN12dEnemyBase_c12UpdateWMClsnER10dBgCh_Actrj(void *thiz, void *clsn, unsigned int a);
+extern void _ZN5dCc_c5ClearEv(void *c);
+extern void _ZN5dCc_c6UpdateEv(void *c);
+extern void _ZN8dActor_c17HugeLandingDustAtER7Vector3b(void *thiz, Vector3 *v, int b);
+extern void _ZN8dActor_c13LandingDustAtER7Vector3b(void *thiz, Vector3 *v, int b);
+extern void _ZN8dActor_c10EarthquakeERK7Vector35Fix12IiE(void *thiz, Vector3 *v, int f);
+void *_ZN5Model8LoadFileER13SharedFilePtr(void *shared);
+void LoadKeyModels(int idx);
+void *_ZN9Animation8LoadFileER13SharedFilePtr(void *shared);
+void *_ZN15TextureSequence8LoadFileER13SharedFilePtr(void *shared);
+void _ZN9ModelBase7SetFileEP8BMD_Fileii(void *mb, void *bmd, int a, int b);
+void _ZN11ShadowModel12InitCylinderEv(void *self);
+void _ZN10dCcAcPos_c4InitEP8dActor_cRK7Vector35Fix12IiES6_jj(
+void *self, void *actor, void *pos, s32 fx, s32 fy, u32 a, u32 b);
+void _ZN18TextureTransformer7PrepareER8BMD_FileR8BTA_File(void *bmd, void *bta);
+void _ZN15TextureSequence7SetFileER8BTP_Filei5Fix12IiEj(void *self, void *f, int a, s32 fix, u32 c);
+void _ZN15MaterialChanger7SetFileER8BMA_Filei5Fix12IiEj(void *self, void *f, int a, s32 fix, u32 c);
+void _ZN10dBgCh_Actr4InitEP8dActor_c5Fix12IiES3_P10Vector3_16S5_(
+void *self, void *actor, s32 fa, s32 fb, void *v0, void *v1);
+extern void *data_ov074_0212292c[];
+extern void *data_ov074_02122948[];
+extern void *data_ov074_02123030;
+extern void *func_021123f4;
+extern void *data_ov074_02122e5c[];
+extern void *func_021124ac;
+extern void _ZN13SharedFilePtr7ReleaseEv(struct SharedFilePtr *self);
+extern void Vec3_Asr(Vector3* d, Vector3* s, int sh);
+extern void Matrix4x3_FromTranslation(void* m, int x, int y, int z);
+extern void Matrix4x3_ApplyInPlaceToTranslation(void* m, int x, int y, int z);
+extern void Matrix4x3_ApplyInPlaceToRotationZXYExt(void* m, int x, int y, int z);
+}
+
+
+/* Forward declarations for every member of this TU: mwccarm lays .text down
+ * in reverse source order, so this file is written descending-ROM and nearly
+ * every intra-TU call is a forward reference. */
+extern "C" {
+void func_ov074_02121300(char* c);
+void func_ov074_02121270(char* r4, char* r8, int r7);
+void func_ov074_02120d74(char *a);
+void func_ov074_02120bb8(char *self, char *p1, char *p2, int mode);
+void func_ov074_02120b90(void *c);
+int func_ov074_02120b24(char* c);
+void func_ov074_0212087c(int* out, char* self, int n_);
+void func_ov074_02120808(char* c);
+int func_ov074_021207b8(struct Obj* self);
+int func_ov074_021206c8(char *c);
+int func_ov074_021204c0(char* c);
+int func_ov074_02120474(char* c);
+void func_ov074_0212042c(char *cc);
+void func_ov074_021203e4(char *cc, int i);
+void func_ov074_021203e0(void);
+void func_ov074_021201f0(char* c);
+void func_ov074_021201ec(void);
+void func_ov074_0212018c(char *c);
+void func_ov074_0212016c(void *c);
+void func_ov074_02120080(char* c);
+void func_ov074_0212007c(void);
+void func_ov074_0211ffcc(char* c);
+void func_ov074_0211ffac(char* c);
+void func_ov074_0211fd74(void* self);
+void func_ov074_0211fd48(char *c);
+void func_ov074_0211fc38(char* c);
+void func_ov074_0211fc34(void);
+int func_ov074_0211fbd0(char* c);
+void func_ov074_0211fb84(char* c);
+void func_ov074_0211fa74(char* c);
+void func_ov074_0211fa08(char* c);
+void func_ov074_0211f5b8(char* c);
+int func_ov074_0211f38c(u8* c);
+void func_ov074_0211f344(void* r6, unsigned char r5);
+void func_ov074_0211f244(dActor_c *self);
+void func_ov074_0211f154(char* c);
+extern int func_ov074_02121a20(void* c, int idx);
+extern int func_ov074_021216f4(void* c);
+extern int func_ov074_02122634(char* self);
+extern int func_ov074_0212229c(int* c);
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 37 -- func_ov074_02121300, 0x02121300, size 0x80 */
+/* -------------------------------------------------------------------------- */
+// @symbol func_ov074_02121300
+/* recovered: shared common types */
+#include "common.h"
+extern "C" {
+
+
+extern void Matrix4x3_ApplyInPlaceToRotationY(void* m, short a);
+extern void Matrix4x3_ApplyInPlaceToScale(void* m, int x, int y, int z);
+extern struct Matrix4x3 data_020a0e68;
+void func_ov074_02121300(char* c){
+  struct Vector3 v;
+  Vec3_Asr(&v, (struct Vector3*)(c+0x5c), 3);
+  Matrix4x3_FromTranslation(&data_020a0e68, v.x, v.y, v.z);
+  Matrix4x3_ApplyInPlaceToRotationY(&data_020a0e68, *(short*)(c+0x8e));
+  Matrix4x3_ApplyInPlaceToScale(&data_020a0e68, *(int*)(c+0x80), *(int*)(c+0x84), *(int*)(c+0x88));
+  *(struct Matrix4x3*)(c+0x37c) = data_020a0e68;
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 36 -- func_ov074_02121270, 0x02121270, size 0x90 */
+/* -------------------------------------------------------------------------- */
+// @symbol func_ov074_02121270
+/* recovered: shared common types */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+void func_ov074_02121270(char* r4, char* r8, int r7){
+  char* dst;
+  *(int*)r4 = 0;
+  *(int*)(r4 + 4) = 0;
+  *(int*)(r4 + 8) = 0;
+  data_020a0e68 = *(struct Matrix4x3*)(r8 + 0x37c);
+  dst = *(char**)(r8 + 0x224) + r7 * 0x30;
+  MulMat4x3Mat4x3(dst, &data_020a0e68, &data_020a0e68);
+  *(int*)r4 = *(int*)((char*)&data_020a0e68 + 0x24);
+  *(int*)(r4 + 4) = *(int*)((char*)&data_020a0e68 + 0x28);
+  *(int*)(r4 + 8) = *(int*)((char*)&data_020a0e68 + 0x2c);
+  Vec3_LslInPlace(r4, 3);
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 35 -- func_ov074_02120d74, 0x02120d74, size 0x4fc */
+/* -------------------------------------------------------------------------- */
+// @symbol func_ov074_02120d74
+/* recovered: shared common types, declarations from a shared header */
+/* recovered: shared common types */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+void func_ov074_02120d74(char *a)
+{
+    struct Vector3 v0;
+    struct Vector3 v1;
+    struct Vector3 v2;
+    struct Vector3 v3;
+    int i;
+    char *player;
+    char *cur;
+    char *p;
+    int hit;
+    int dir;
+    int dir2;
+    u32 s;
+    u32 flags;
+    u16 type;
+    int b;
+    int b2;
+
+    DecIfAbove0_Byte((u8 *)(a + 0x605));
+    s = *(u32 *)(a + 0x5cc);
+    if (s == 1)
+        return;
+    if (s == 2)
+        return;
+    if (s - 5 <= 3)
+        return;
+
+    player = 0;
+    i = 0;
+    p = a + 0x110;
+    do {
+        cur = p;
+        if (*(u32 *)(cur + 0x24) != 0)
+            player = _ZN8dActor_c10FindWithIDEj(*(u32 *)(cur + 0x24));
+        if (player != 0)
+            break;
+        i++;
+        p += 0x40;
+    } while (i < 4);
+    if (player == 0)
+        return;
+
+    flags = *(u32 *)(cur + 0x20);
+    hit = 0;
+    if (flags & 0x2000) {
+        hit = 1;
+    } else {
+        type = *(u16 *)(player + 0xc);
+        b = (int)(type == 0xbf);
+        if (b != 0) {
+            if (*(u8 *)(a + 0x605) == 0) {
+                if (!(flags & 0x26fe0) && *(u8 *)(player + 0x708) == 0) {
+                    if (cur == a + 0x190) {
+                        if (*(s16 *)(a + 0x5f6) != 0) {
+                            _ZN6Player12Unk_020c6a10Ej(player, 2);
+                            *(u8 *)(a + 0x605) = 0x1e;
+                        } else {
+                            v1.x = *(int *)(a + 0x5c);
+                            v1.y = *(int *)(a + 0x60);
+                            v1.z = *(int *)(a + 0x64);
+                            _ZN6Player4HurtERK7Vector3j5Fix12IiEjjj(player, &v1, 1, 0x5000, 1, 0, 1);
+                        }
+                    } else if (cur == a + 0x1d0) {
+                        if (*(s16 *)(a + 0x5f6) != 0) {
+                            _ZN6Player12Unk_020c6a10Ej(player, 2);
+                            *(u8 *)(a + 0x605) = 0x1e;
+                        } else {
+                            v2.x = *(int *)(a + 0x5c);
+                            v2.y = *(int *)(a + 0x60);
+                            v2.z = *(int *)(a + 0x64);
+                            _ZN6Player4HurtERK7Vector3j5Fix12IiEjjj(player, &v2, 1, 0x5000, 1, 0, 1);
+                        }
+                    } else {
+                        v3.x = *(int *)(a + 0x5c);
+                        v3.y = *(int *)(a + 0x60);
+                        v3.z = *(int *)(a + 0x64);
+                        _ZN6Player4HurtERK7Vector3j5Fix12IiEjjj(player, &v3, 1, 0x5000, 1, 0, 1);
+                        *(u8 *)(a + 0x605) = 0x10;
+                    }
+                }
+                if (*(u8 *)(player + 0x703) != 0) {
+                    if (*(u32 *)(cur + 0x20) & 0x3c0) {
+                        hit = 1;
+                    } else {
+                        int ang = Vec3_HorzAngle((struct Vector3 *)(a + 0x5c), (struct Vector3 *)(player + 0x5c));
+                        dir = 1;
+                        if (AngleDiff(ang, *(s16 *)(a + 0x8e)) > 0x4000)
+                            dir = -1;
+                        func_ov074_02121a4c(a, 7);
+                        func_ov074_021203e4(a, 8);
+                        func_ov074_02120bb8(a, player, cur, 0);
+                        func_02012694(0x15a, a + 0x74);
+                        *(s16 *)(a + 0x5f6) = dir * (*(int *)(a + 0x5f0) * -0xc8);
+                    }
+                } else {
+                    if (cur == a + 0x150) {
+                        if (*(u32 *)(cur + 0x20) & 0x80) {
+                            if (*(u32 *)(a + 0x5cc) == 4)
+                                func_ov074_021203e4(a, 2);
+                            else
+                                func_ov074_021203e4(a, 4);
+                            if (*(int *)(player + 8) == 0) {
+                                v0.x = *(int *)(a + 0x5c);
+                                v0.y = *(int *)(a + 0x60);
+                                v0.z = *(int *)(a + 0x64);
+                                if (*(u32 *)(a + 0x5dc) == 0 && _ZN8dActor_c10FindWithIDEj(*(u32 *)(a + 0x5dc)) == 0) {
+                                    int k = ((u16)(s16)(*(int *)(a + 0x5f0) * ((s16)((*(int *)(a + 0x80) / 0x1000) * 0x350 + 0x500) + 0x600) + (*(s16 *)(a + 0x5f4) + *(s16 *)(a + 0x5f8)))) >> 4;
+                                    v0.x = (int)(((long long)data_02082214[k * 2] * 0x546000LL + 0x800) >> 12);
+                                    v0.y += 0x32000;
+                                    v0.z = (int)(((long long)data_02082214[k * 2 + 1] * 0x546000LL + 0x800) >> 12);
+                                    *(int *)(a + 0x5dc) = *(int *)(_ZN8dActor_c5SpawnEjjRK7Vector3PK10Vector3_16as(0x115, 0, &v0, 0, *(s8 *)(a + 0xcc), -1) + 4);
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+        } else {
+            b2 = (int)(type == 0xc8);
+            if (b2 != 0) {
+                if (flags & 0x22000)
+                    hit = 1;
+            }
+        }
+    }
+    if (hit == 0)
+        return;
+    func_02012694(0x15b, a + 0x74);
+    {
+    int ang2 = Vec3_HorzAngle((struct Vector3 *)(a + 0x5c), (struct Vector3 *)(player + 0x5c));
+    dir2 = 1;
+    if (AngleDiff(ang2, *(s16 *)(a + 0x8e)) > 0x4000) {
+        func_ov074_02121a4c(a, dir2);
+        dir2 = -1;
+    } else {
+        func_ov074_02121a4c(a, 2);
+    }
+    }
+    *(s16 *)(a + 0x5f6) = dir2 * (*(int *)(a + 0x5f0) * -0x190);
+    func_ov074_021203e4(a, 5);
+    func_ov074_02120bb8(a, player, cur, 1);
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 34 -- func_ov074_02120bb8, 0x02120bb8, size 0x1bc */
+/* -------------------------------------------------------------------------- */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+void func_ov074_02120bb8(char *self, char *p1, char *p2, int mode)
+{
+    Vector3 v, dst, src, delta;
+    s32 aX, aY;
+    s32 iX, iY;
+    s32 s, w;
+
+    dst.x = *(s32*)(self + 0x5c);
+    dst.y = *(s32*)(self + 0x60);
+    dst.z = *(s32*)(self + 0x64);
+    {
+        Vector3 *pv = (Vector3*)(p1 + 0x5c);
+        src.x = pv->x;
+        src.y = pv->y;
+        src.z = pv->z;
+    }
+    {
+        s32 dy = *(s32*)(p2 + 8);
+        w = *(s32*)(p2 + 4);
+        dst.y += dy;
+    }
+    src.y += 0x46000;
+    Vec3_Sub(&delta, &src, &dst);
+
+    v.x = delta.x;
+    v.y = delta.y;
+    v.z = delta.z;
+    aX = _ZN4cstd5atan2E5Fix12IiES1_(v.x, v.z);
+    aY = _ZN4cstd5atan2E5Fix12IiES1_(v.y, Vec3_HorzLen(&v));
+
+    iY = (u16)aY >> 4;
+    iX = (u16)aX >> 4;
+
+    s = (s32)(((long long)w * data_02082214[iY * 2 + 1] + 0x800) >> 12);
+    dst.x += (s32)(((long long)s * data_02082214[iX * 2] + 0x800) >> 12);
+    dst.y += (s32)(((long long)w * data_02082214[iY * 2] + 0x800) >> 12);
+    dst.z += (s32)(((long long)s * data_02082214[iX * 2 + 1] + 0x800) >> 12);
+
+    switch (mode) {
+    case 0:
+        _ZN8Particle6System9NewSimpleEj5Fix12IiES2_S2_(0xf6, dst.x, dst.y, dst.z);
+        break;
+    case 1:
+        _ZN8Particle6System9NewSimpleEj5Fix12IiES2_S2_(0x43, dst.x, dst.y, dst.z);
+        _ZN8Particle6System9NewSimpleEj5Fix12IiES2_S2_(0x44, dst.x, dst.y, dst.z);
+        break;
+    }
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 33 -- func_ov074_02120b90, 0x02120b90, size 0x28 */
+/* -------------------------------------------------------------------------- */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+void func_ov074_02120b90(void *c) {
+    int idx = *(unsigned char*)((char*)c + 0x604);
+    int r3 = *(int*)((char*)c + 0x5f0);
+    unsigned short r1 = data_ov074_02122dfc[idx];
+    *(short*)((char*)c + 0x5f8) = r3 * r1;
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 32 -- func_ov074_02120b24, 0x02120b24, size 0x6c */
+/* -------------------------------------------------------------------------- */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+int func_ov074_02120b24(char* c)
+{
+    int num = *(short*)(c + 0x5f6) << 12;
+    int den;
+    int r;
+    if (num < 0) num = -num;
+    den = data_ov074_0212291c[*(unsigned char*)(c + 0x604)] + (data_ov074_02122dfc[3] << 12) + 1;
+    r = _ZN4cstd4fdivEii(num, den);
+    if (r < 0x1000) return 0x1000;
+    if (r > 0x3000) return 0x3000;
+    return r;
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 31 -- func_ov074_0212087c, 0x0212087c, size 0x2a8 */
+/* -------------------------------------------------------------------------- */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+void func_ov074_0212087c(int* out, char* self, int n_)
+{
+    int n;
+
+    n = LI(n_);
+
+    out[0] = *(int*)(self + 0x5c);
+    out[1] = *(int*)(self + 0x60);
+    out[2] = *(int*)(self + 0x64);
+
+    if (*(int*)(_ZN8dActor_c13ClosestPlayerEv(self) + 8) == 3) {
+        int rem;
+        int quo;
+        int scale;
+        int q;
+        int sum;
+        int prod;
+        int diff;
+        int idx;
+        u8 mode;
+
+        rem = n % 3;
+        quo = LI(n) / 3;
+        mode = *(u8*)(self + 0x604);
+        if (mode == 3) {
+            scale = 0x546000;
+        } else if (mode == 2) {
+            scale = quo * 0xb4000 + 0x4ec000;
+        } else {
+            scale = quo * 0xb4000 + 0x438000;
+        }
+
+        q = *(int*)(self + 0x80) / 0x1000;
+        sum = (s16)(q * 0x350 + 0x500) + (s16)(rem * 0x600);
+        prod = *(int*)(self + 0x5f0) * sum;
+        diff = (*(s16*)(self + 0x5f4) + *(s16*)(self + 0x5f8)) - prod;
+        idx = ((u16)(s16)diff >> 4) * 2;
+
+        out[0] = (int)(((s64)scale * data_02082214[idx] + 0x800) >> 12);
+        *(int*)AT(out, 4) += 0x32000;
+        out[2] = (int)(((s64)scale * data_02082214[idx + 1] + 0x800) >> 12);
+    } else {
+        int lim;
+        int one;
+        int sign;
+        int rem2;
+        int scale2;
+        int q2;
+        int sum2;
+        int prod2;
+        int prod3;
+        int diff2;
+        int idx2;
+
+        lim = 3;
+        if (*(u8*)(self + 0x604) == 1)
+            lim = 5;
+
+        if (n < lim) {
+            one = 1;
+            sign = -1;
+            rem2 = n % lim;
+            scale2 = 0x4e2000 + (rem2 * 0x64 << 12);
+        } else {
+            one = 1;
+            sign = one;
+            rem2 = n % lim;
+            scale2 = 0x4e2000 + (rem2 * 0x64 << 12);
+        }
+
+        q2 = *(int*)(self + 0x80) / 0x1000;
+        sum2 = (s16)(q2 * 0x350 + 0x250) + (s16)(one * 0x600);
+        prod2 = *(int*)(self + 0x5f0) * sum2;
+        prod3 = sign * prod2;
+        diff2 = (*(s16*)(self + 0x5f4) + *(s16*)(self + 0x5f6)) - prod3;
+        idx2 = ((u16)(s16)diff2 >> 4) * 2;
+
+        out[0] = (int)(((s64)scale2 * data_02082214[idx2] + 0x800) >> 12);
+        *(int*)AT(out, 4) += 0x32000;
+        out[2] = (int)(((s64)scale2 * data_02082214[idx2 + 1] + 0x800) >> 12);
+    }
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 30 -- func_ov074_02120808, 0x02120808, size 0x74 */
+/* -------------------------------------------------------------------------- */
+extern "C" void func_ov074_02120808(char* c)
+{
+    int ang = (short)(*(short*)(c + 0x5f4) + *(int*)(c + 0x5f0) * 0x4500);
+    if (AngleDiff(ang, _ZN8dActor_c18HorzAngleToCPlayerEv(c)) <= 0x4000)
+        return;
+    func_ov074_02121a4c(c, 6);
+    *(int *)(((int)(c + 0x5f0))) *= -1;
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 29 -- func_ov074_021207b8, 0x021207b8, size 0x50 */
+/* -------------------------------------------------------------------------- */
+/* func_ov074_021207b8 — if func_ov074_02121a20(this, 6) and the animation at
+ * this+0x260 has finished, flip the angle at this+0x8e by 0x8000 and return 1;
+ * else return 0.
+ * Callees: func_ov074_02121a20, _ZN9Animation8FinishedEv.
+ */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+int func_ov074_021207b8(struct Obj* self)
+{
+    if (func_ov074_02121a20(self, 6)) {
+        if (_ZN9Animation8FinishedEv(&self->anim)) {
+            *(s16*)((char*)self + 0x8e) += 0x8000;
+            return 1;
+        }
+    }
+    return 0;
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 28 -- func_ov074_021206c8, 0x021206c8, size 0xf0 */
+/* -------------------------------------------------------------------------- */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+int func_ov074_021206c8(char *c) {
+    if (func_ov074_02121a20(c, 8) != 0) {
+        if (ApproachAngle((short *)(c + 0x5f6), 0, 0x1e, 0x20, 1) == 0) {
+            if ((unsigned short)(*(int *)(c + 0x268) >> 12) == 0) {
+                func_ov074_02121a4c(c, 9);
+            }
+            goto ret0;
+        }
+        func_ov074_021216f4(c);
+        goto ret0;
+    }
+    if (func_ov074_02121a20(c, 0xa) != 0) {
+        func_ov074_02121a4c(c, 9);
+        goto ret0;
+    }
+    if (func_ov074_02121a20(c, 9) != 0) {
+        if (_ZN9Animation8FinishedEv(c + 0x260) != 0) {
+            return 1;
+        }
+        goto ret0;
+    }
+    return 1;
+ret0:
+    return 0;
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 27 -- func_ov074_021204c0, 0x021204c0, size 0x208 */
+/* -------------------------------------------------------------------------- */
+// @symbol func_ov074_021204c0
+/* recovered: Goomboss (daKuriKing_c) shrink/grow interpolation tick.
+ *
+ * Reached from func_ov074_0211f860 and func_ov074_0211fb44. Picks a pair of
+ * float sizes out of data_ov074_02122e4c (12288.0f, 9557.0f, 6826.0f,
+ * 4096.0f -- Fix12 scales stored as IEEE singles), divides their difference
+ * by the remaining step count in f_607, and steps f_5ec toward the next one.
+ * Math_Function_0203b14c eases f_5e4 toward that target; while it is still
+ * moving, the scale triple at f_80/f_84/f_88 is driven from a sin/cos pair
+ * out of data_02082214 indexed by the f_5fa phase, which advances 0x1230 per
+ * frame. Every step below the last also spits a 0xb1 particle a fixed Fix12
+ * offset above the boss.
+ *
+ * The float arithmetic goes through the ITCM soft-float block by hand, the
+ * way the rest of this tree calls it: func_01ffa594 (single add/sub),
+ * func_01ffa344 (float to int), func_01ffa4bc (int to float) and
+ * func_01ff9378, with cstd::fdiv doing the Fix12 divide in between.
+ *
+ * Shape notes: the particle position temp must stay MEMORY-HOMED -- reading
+ * its words back through the address (the ((int *)&v)[n] form) is what keeps
+ * the three dead stores the ROM emits; plain member reads scalarize the
+ * struct away and cost four instructions. And k has to be computed BEFORE d,
+ * which is worth ten words of register identity in the tail: with d first,
+ * the index web takes r2 and the f_5e4 web takes r3, exactly reversed from
+ * the ROM.
+ *
+ * Matched byte-for-byte with mwccarm 2004/b56 (ov074); linkcheck VERIFIED.
+ */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+int func_ov074_021204c0(char* c) {
+    struct Vector3 v;
+    int d, k;
+
+    *(int*)(c + 0x5ec) = func_01ffa344(func_01ff9378(
+        (int)data_ov074_02122e4c[*(unsigned char*)(c + 0x604) + 1],
+        func_01ffa4bc(_ZN4cstd4fdivEii(
+            func_01ffa344(func_01ffa594((int)data_ov074_02122e4c[*(unsigned char*)(c + 0x604)],
+                                        (int)data_ov074_02122e4c[*(unsigned char*)(c + 0x604) + 1])),
+            *(unsigned char*)(c + 0x607) << 12))));
+
+    if (Math_Function_0203b14c((int*)(c + 0x5e4), *(int*)(c + 0x5ec), 0x78, 0x100, 0x20) == 0) {
+        if (*(short*)(c + 0x5fa) == 0) {
+            *(int*)(c + 0x80) = *(int*)(c + 0x5e4);
+            *(int*)(c + 0x84) = *(int*)(c + 0x5e4);
+            *(int*)(c + 0x88) = *(int*)(c + 0x5e4);
+        }
+        if (*(unsigned char*)(c + 0x607) <= 1) return 1;
+        (*(unsigned char*)(c + 0x607))--;
+        if (*(int*)(c + 0x5cc) == 6) {
+            v.x = *(int*)(c + 0x5c);
+            v.y = *(int*)(c + 0x60);
+            v.z = *(int*)(c + 0x64);
+            v.y += (int)(((long long)*(int*)(c + 0x5e4) * (long long)0x190000 + 0x800) >> 12);
+            _ZN8Particle6System9NewSimpleEj5Fix12IiES2_S2_(0xb1, ((int*)&v)[0], ((int*)&v)[1], ((int*)&v)[2]);
+        }
+    }
+    k = (*(unsigned short*)(c + 0x5fa) >> 4) * 2;
+    d = *(int*)(c + 0x5ec) - *(int*)(c + 0x5e4) + 0x11e;
+    *(int*)(c + 0x84) = *(int*)(c + 0x5e4) + (int)(((long long)d * data_02082214[k] + 0x800) >> 12);
+    *(int*)(c + 0x80) = *(int*)(c + 0x5e4);
+    *(int*)(c + 0x80) = *(int*)(c + 0x5e4) + (int)(((long long)d * data_02082214[k + 1] + 0x800) >> 12);
+    *(int*)(c + 0x88) = *(int*)(c + 0x80);
+    *(short*)(c + 0x5fa) += 0x1230;
+    return 0;
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 26 -- func_ov074_02120474, 0x02120474, size 0x4c */
+/* -------------------------------------------------------------------------- */
+extern "C" {
+int func_ov074_02120474(char* c) {
+    _ZN9Animation7AdvanceEv(c + 0x3d0);
+    unsigned char idx = *(unsigned char*)(c + 0x604);
+    unsigned int val = *(unsigned int*)(c + 0x3d8);
+    unsigned short tbl = data_ov074_02122e04[idx];
+    unsigned int v = (val << 4) >> 0x10;
+    if (v >= tbl) {
+        *(unsigned int*)(c + 0x3d8) = (unsigned int)tbl << 0xc;
+        return 1;
+    }
+    return 0;
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 25 -- func_ov074_0212042c, 0x0212042c, size 0x48 */
+/* -------------------------------------------------------------------------- */
+struct C; typedef void (C::*PMF)();
+struct PmfEntry { PMF pmf[2]; };
+extern PmfEntry data_ov074_021230f8[];
+struct C { char pad[0x5cc]; int idx; };
+extern "C" void func_ov074_0212042c(char *cc) { C *c = (C *)cc; int j = c->idx; (c->*data_ov074_021230f8[j].pmf[1])(); }
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 24 -- func_ov074_021203e4, 0x021203e4, size 0x48 */
+/* -------------------------------------------------------------------------- */
+extern "C" void func_ov074_021203e4(char *cc, int i) { C *c = (C *)cc; c->idx = i; int j = c->idx; (c->*data_ov074_021230f8[j].pmf[0])(); }
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 23 -- func_ov074_021203e0, 0x021203e0, size 0x4 */
+/* -------------------------------------------------------------------------- */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+void func_ov074_021203e0(void)
+{
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 22 -- func_ov074_021201f0, 0x021201f0, size 0x1f0 */
+/* -------------------------------------------------------------------------- */
+// @symbol func_ov074_021201f0
+/* recovered: Goomboss (daKuriKing_c) state-0 tick -- the intro cutscene.
+ *
+ * Frames the boss for the opening conversation: pins the camera to a fixed
+ * look-at/eye pair derived from the boss position, turns the closest player
+ * around to face him (angle + 0x7fff), snaps the boss's own facing to
+ * f_5f0 * 0x4500 + f_5f4, then runs the talk handshake. Player::StartTalk
+ * and Player::ShowMessage each gate the rest of the tick, so a refused talk
+ * simply retries next frame; on success the music layer flips once (guarded
+ * by the f_60c latch) and the state advances to 1.
+ *
+ * Shape notes: the angle temp is a Vector3_16 STRUCT COPY, not three field
+ * assignments -- mwccarm scalarizes the field form away and drops the
+ * memory home the ROM keeps at sp+8. msgpos[0] is filled through a base
+ * pointer (the ov073 idiom) while msgpos[1] uses direct offsets; that
+ * difference is in the ROM, not an accident. msgpos[0] is dead, kept alive
+ * only because &msgpos[1] homes the whole array.
+ *
+ * Matched byte-for-byte with mwccarm 2004/b56 (ov074); linkcheck VERIFIED.
+ */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+void func_ov074_021201f0(char* c) {
+    struct Ang16 { short v[3]; };
+    struct Ang16 ang;
+    struct Vector3 look, pos;
+    struct Vector3 msgpos[2];
+    char* player;
+    void* cam;
+    short msg;
+
+    player = _ZN8dActor_c13ClosestPlayerEv(c);
+    cam = data_0209f318;
+    _ZN6Camera9SetFlag_3Ev(cam);
+
+    look.x = *(int*)(c + 0x5c);
+    look.y = *(int*)(c + 0x60);
+    look.z = *(int*)(c + 0x64);
+    pos.x = *(int*)(c + 0x5c);
+    pos.y = *(int*)(c + 0x60);
+    pos.z = *(int*)(c + 0x64);
+    look.x += 0x130000;
+    look.y = 0x1cd000;
+    look.z -= 0x2a0000;
+    pos.x -= 0x2f0000;
+    pos.y += 0x20000;
+    pos.z += 0x254000;
+    _ZN6Camera9SetLookAtERK7Vector3(cam, &look);
+    _ZN6Camera6SetPosERK7Vector3(cam, &pos);
+
+    ang = *(struct Ang16*)(c + 0x8c);
+    ang.v[1] += 0x7fff;
+    *(short*)(player + 0x8c) = ang.v[0];
+    *(short*)(player + 0x8e) = ang.v[1];
+    *(short*)(player + 0x90) = ang.v[2];
+
+    {
+        int pv = (int)(player + 0x5c);
+        msgpos[0].x = *(int*)pv;
+        msgpos[0].y = *(int*)(pv + 4);
+        msgpos[0].z = *(int*)(pv + 8);
+    }
+    msgpos[1].x = *(int*)(c + 0x5c);
+    msgpos[1].y = *(int*)(c + 0x60);
+    msgpos[1].z = *(int*)(c + 0x64);
+    msgpos[1].y += 0x78000;
+
+    *(short*)(c + 0x8e) = *(int*)(c + 0x5f0) * 0x4500 + *(short*)(c + 0x5f4);
+    *(char**)(c + 0x5c8) = player;
+    if (_ZN6Player9StartTalkER7fBase_cb(player, c, 1) == 0) return;
+    _ZN7Message11PrepareTalkEv();
+    if (*(unsigned char*)(c + 0x60c) == 0) {
+        _ZN5Sound22LoadAndSetMusic_Layer3Ej(0x2c);
+        *(unsigned char*)(c + 0x60c) = 1;
+    }
+    msg = (unsigned short)(*(int*)(player + 8) + 0xd3);
+    if (_ZN6Player11ShowMessageER7fBase_cjPK7Vector3hh(player, c, msg, &msgpos[1], 0, 2) == 0) return;
+    func_02012694(0x15a, c + 0x74);
+    func_ov074_021203e4(c, 1);
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 21 -- func_ov074_021201ec, 0x021201ec, size 0x4 */
+/* -------------------------------------------------------------------------- */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+void func_ov074_021201ec(void)
+{
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 20 -- func_ov074_0212018c, 0x0212018c, size 0x60 */
+/* -------------------------------------------------------------------------- */
+extern "C" {
+extern void func_02011d20(void);
+extern char *data_0209f318;
+void func_ov074_0212018c(char *c)
+{
+    if (_ZN6Player12GetTalkStateEv(*(void **)(c + 0x5c8)) != -1)
+        return;
+    (*(int *)(((int)data_0209f318 + 0x154))) &= ~8;
+    _ZN5Sound22LoadAndSetMusic_Layer3Ej(0x2d);
+    func_02011d20();
+    _ZN7Message7EndTalkEv();
+    func_ov074_021203e4(c, 2);
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 19 -- func_ov074_0212016c, 0x0212016c, size 0x20 */
+/* -------------------------------------------------------------------------- */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+void func_ov074_0212016c(void *c) {
+    func_ov074_02121a4c((char*)c, 0xa);
+    func_ov074_02120b90(c);
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 18 -- func_ov074_02120080, 0x02120080, size 0xec */
+/* -------------------------------------------------------------------------- */
+extern "C" {
+void func_ov074_02120080(char* c){
+  if(*(unsigned char*)(c+0x602) == 0){
+    func_ov074_021203e4(c, 2);
+    return;
+  }
+  if(func_ov074_02121a20(c, 0xa) != 0){
+    if(_ZN9Animation8FinishedEv(c+0x260) == 0) return;
+    func_ov074_02121a4c(c, 8);
+    return;
+  }
+  ApproachAngle(c+0x5f6, *(short*)(c+0x500+0xf8), 0xa, 0x20, 1);
+  func_ov074_021216f4(c);
+  if(_ZN8dActor_c13DistToCPlayerEv(c) >= 0x4b0000) return;
+  if(AngleDiff(_ZN8dActor_c18HorzAngleToCPlayerEv(c), *(short*)(c+0x8e)) <= 0x4000) return;
+  func_ov074_021203e4(c, 4);
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 17 -- func_ov074_0212007c, 0x0212007c, size 0x4 */
+/* -------------------------------------------------------------------------- */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+void func_ov074_0212007c(void)
+{
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 16 -- func_ov074_0211ffcc, 0x0211ffcc, size 0xb0 */
+/* -------------------------------------------------------------------------- */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+void func_ov074_0211ffcc(char* c) {
+  if (func_ov074_02121a20(c, 5) == 0) {
+    if (func_ov074_021206c8(c) == 0) return;
+    func_ov074_02121a4c(c, 5);
+    return;
+  }
+  if (_ZN8dActor_c13DistToCPlayerEv(c) >= 0x4b0000) {
+    func_ov074_021203e4(c, 3);
+  } else {
+    if (AngleDiff(_ZN8dActor_c18HorzAngleToCPlayerEv(c), *(s16*)(c+0x8e)) < 0x3000) {
+      func_ov074_021203e4(c, 3);
+      return;
+    }
+  }
+  if (*(unsigned char*)(c+0x602) != 0) return;
+  func_ov074_021203e4(c, 2);
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 15 -- func_ov074_0211ffac, 0x0211ffac, size 0x20 */
+/* -------------------------------------------------------------------------- */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+void func_ov074_0211ffac(char* c) {
+    func_ov074_02120808(c);
+    *(unsigned short *)(c + 0x500 + 0xfc) = 0;
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 14 -- func_ov074_0211fd74, 0x0211fd74, size 0x238 */
+/* -------------------------------------------------------------------------- */
+// @symbol func_ov074_0211fd74
+/* recovered: shared common types, declarations from a shared header */
+/* recovered: shared common types */
+extern "C" void func_ov074_0211fd74(void* self)
+{
+    u8* c = (u8*)self;
+
+    if (func_ov074_02121a20(self, 6) != 0) {
+        if (func_ov074_021207b8((struct Obj*)self) == 0) return;
+        func_ov074_02121a4c((char*)self, 0xb);
+        return;
+    } else {
+        if (func_ov074_02121a20(self, 0xb) != 0) goto L90;
+        if (func_ov074_021206c8((char*)self) == 0) return;
+        func_ov074_02121a4c((char*)self, 0xb);
+        return;
+    }
+
+L90:;
+    int r5 = data_ov074_02122d80[*(u8*)(c+0x604)];
+    u8* pl = (u8*)_ZN8dActor_c13ClosestPlayerEv(self);
+    if (*(u8*)(c+0x604) == 1 && *(s32*)(pl+8) != 3) r5 = r5 + 1;
+
+    if (_ZN9Animation8FinishedEv((void*)(c+0x260)) == 0) return;
+    if (DecIfAbove0_Short((u16*)(c+0x5fc)) != 0) return;
+
+    Vector3 pos;
+    func_ov074_0212087c((int*)&pos, (char*)self, *(u8*)(c+0x602));
+    Vector3 pp;
+    pp.x = pos.x;
+    pp.y = pos.y;
+    pp.z = pos.z;
+    _ZN8dActor_c19DisappearPoofDustAtERK7Vector3(self, &pp);
+
+    u32 sid = 0xeeee;
+    if (*(s32*)(pl+8) != 3) {
+        if (*(u8*)(c+0x602) % 2 == 0) sid = 0xeeef;
+    }
+
+    u8* sp2 = (u8*)_ZN8dActor_c5SpawnEjjRK7Vector3PK10Vector3_16as(
+        0xc8, sid, &pos, (const struct Vector3_16*)(c+0x8c), *(signed char*)(c+0xcc), -1);
+    *(s32*)(sp2+0x43c) = *(s32*)(c+4);
+    Vector3 v2;
+    v2.x = *(s32*)(c+0x5d0);
+    v2.y = *(s32*)(c+0x5d4);
+    v2.z = *(s32*)(c+0x5d8);
+    func_ov084_021296b0(sp2, &v2);
+
+    *(u8*)(sp2+0x474) = *(u8*)(c+0x602);
+    u8* p602 = (u8*)(((int)c + 0x602));
+    *p602 = (u8)(*p602 + 1);
+    *(u16*)((c+0x500)+0xfc) = 2;
+
+    u8* pl2 = (u8*)_ZN8dActor_c13ClosestPlayerEv(self);
+    int dv = 3;
+    if (*(s32*)(pl2+8) != 3) {
+        if (*(u8*)(c+0x604) == 1) dv = 5;
+    }
+    if (*(u8*)(c+0x602) % dv == 0) *(s32*)(c+0x268) = 0;
+
+    if (*(u8*)(c+0x602) < r5) return;
+    func_ov074_021203e4((char*)self, 3);
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 13 -- func_ov074_0211fd48, 0x0211fd48, size 0x2c */
+/* -------------------------------------------------------------------------- */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+void func_ov074_0211fd48(char *c)
+{
+    unsigned char value;
+
+    (*(unsigned char *)(((int)c + 0x604)))--;
+    value = *(unsigned char *)(c + 0x604);
+    if (value > 5)
+        value = 5;
+    *(unsigned char *)(c + 0x604) = value;
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 12 -- func_ov074_0211fc38, 0x0211fc38, size 0x110 */
+/* -------------------------------------------------------------------------- */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+void func_ov074_0211fc38(char* c)
+{
+    if (ApproachAngle(c + 0x5f6, 0, 0xa, 0x100, 1) != 0) goto reset;
+    if (_ZN9Animation8FinishedEv(c + 0x260) == 0) return;
+    if (*(unsigned char*)(c + 0x604) != 0) goto detox;
+    {
+        void* player = _ZN8dActor_c13ClosestPlayerEv(c);
+        if (_ZN6Player9StartTalkER7fBase_cb(player, c, 1) == 0) return;
+        _ZN7Message11PrepareTalkEv();
+        if (func_ov074_02121a20(c, 1) != 0 || func_ov074_02121a20(c, 2) != 0) {
+            func_ov074_02121a4c(c, 0xc);
+            return;
+        }
+        func_ov074_021203e4(c, 7);
+        *(void**)(c + 0x5c8) = player;
+        return;
+    }
+detox:
+    func_ov074_021203e4(c, 6);
+    return;
+reset:
+    func_ov074_021216f4(c);
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 11 -- func_ov074_0211fc34, 0x0211fc34, size 0x4 */
+/* -------------------------------------------------------------------------- */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+void func_ov074_0211fc34(void)
+{
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 10 -- func_ov074_0211fbd0, 0x0211fbd0, size 0x64 */
+/* -------------------------------------------------------------------------- */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+int func_ov074_0211fbd0(char* c){
+    if(ApproachAngle((char*)c+0x5f6, 0, 0xa, 0x100, 1)==0)
+        return ((int (*)(char*, int))func_ov074_021203e4)(c, 3);
+    return func_ov074_021216f4(c);
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 9 -- func_ov074_0211fb84, 0x0211fb84, size 0x4c */
+/* -------------------------------------------------------------------------- */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+void func_ov074_0211fb84(char* c){
+  func_ov074_02120808(c);
+  char* b=c+0x500;
+  *(short*)(b+0xfa)=0;
+  *(char*)(c+0x603)=0;
+  unsigned short t=3-(unsigned char)c[0x604];
+  if(t>2)t=2;
+  *(int*)(c+0x3ec)=t<<12;
+  *(char*)(c+0x607)=3;
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 8 -- func_ov074_0211fa74, 0x0211fa74, size 0x110 */
+/* -------------------------------------------------------------------------- */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+void func_ov074_0211fa74(char* c)
+{
+    char* a = c;
+    int b;
+    switch (*(unsigned char*)(a + 0x603)) {
+    case 0:
+        if (func_ov074_02121a20(a, 6) != 0) {
+            unsigned char* p;
+            b = func_ov074_021207b8((struct Obj*)a);
+            if (b == 0) return;
+            p = (unsigned char*)(((int)a + 0x603));
+            *p = *p + 1;
+            func_ov074_02121a4c(a, 0);
+            func_02012694(0x15c, a + 0x74);
+            return;
+        }
+        {
+            unsigned char* p = (unsigned char*)(((int)a + 0x603));
+            *p = *p + 1;
+        }
+        func_ov074_02121a4c(a, 0);
+        func_02012694(0x15c, a + 0x74);
+        return;
+    case 1:
+        {
+            b = func_ov074_02120474(a);
+            if (func_ov074_021204c0(a) == 0) return;
+            if (b == 0) return;
+            func_ov074_021203e4(a, 2);
+        }
+        return;
+    }
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 7 -- func_ov074_0211fa08, 0x0211fa08, size 0x6c */
+/* -------------------------------------------------------------------------- */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+void func_ov074_0211fa08(char* c){
+  *(char*)(c+0x604)=0;
+  int v=func_01ffa344((int)data_ov074_02122e4c[1]);
+  *(int*)(c+0x80)=v;
+  *(int*)(c+0x84)=v;
+  *(int*)(c+0x88)=v;
+  *(int*)(c+0x3ec)=0x2000;
+  *(int*)(c+0x3d8)=((int)data_ov074_02122e04[1])<<0xc;
+  *(char*)(c+0x603)=0;
+  ((void (*)(char*, int, int))func_ov074_02121a4c)(c, 3, 0);
+  *(char*)(c+0x607)=3;
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 6 -- func_ov074_0211f5b8, 0x0211f5b8, size 0x450 */
+/* -------------------------------------------------------------------------- */
+extern "C" void func_ov074_0211f5b8(char* c)
+{
+    Vector3 vzero;
+    Vector3 vmsg;
+    void* player;
+    s32 state;
+    void* data0209f318;
+
+    vzero.x = 0;
+    vzero.y = 0;
+    vzero.z = 0;
+    vmsg.x = *(s32*)(c + 0x5c);
+    vmsg.y = *(s32*)(c + 0x60);
+    vmsg.z = *(s32*)(c + 0x64);
+    vmsg.y = vmsg.y + 0x78000;
+    data0209f318 = data_0209f318;
+    player = *(void**)(c + 0x5c8);
+
+    if (_ZN9Animation8FinishedEv(c + 0x260) != 0) {
+        func_ov074_02121a4c(c, 4);
+    }
+
+    switch (*(u8*)(c + 0x603)) {
+    case 0:
+        func_ov074_0211f154(c);
+        {
+            u16 msgId = (u16)(*(s32*)((char*)player + 8) + 0xd7);
+            state = _ZN6Player12GetTalkStateEv(player);
+            if (state != 0) {
+                if (state == 1)
+                    return;
+                if (state == 2)
+                    goto inc603;
+                return;
+            }
+            if (_ZN6Player11ShowMessageER7fBase_cjPK7Vector3hh(player, c, (s16)msgId, &vmsg, 1, 2) == 0)
+                return;
+            func_0201277c(0x15a);
+        }
+        return;
+    inc603:
+        *(u8*)AT(c, 0x603) += 1;
+        return;
+    case 1:
+        if (DecIfAbove0_Short((u16*)(c + 0x5fc)) == 0) {
+            void* actor = _ZN8dActor_c15FindWithActorIDEjPS_(0xc8, 0);
+            if (actor != 0) {
+                u32 id = 0xc8;
+                while (1) {
+                    s32 flag = (*(u32*)((char*)actor + 0xb0) & 0x40000) ? 1 : 0;
+                    if (flag == 0) {
+                        Vector3 v2;
+                        s32* src = (s32*)AT(actor, 0x5c);
+                        v2.x = src[0];
+                        v2.y = src[1];
+                        v2.z = src[2];
+                        _ZN8dActor_c10PoofDustAtERK7Vector3(c, &v2);
+                        _ZN7fBase_c18MarkForDestructionEv(actor);
+                        *(u16*)(c + 0x5fc) = 2;
+                        return;
+                    }
+                    actor = _ZN8dActor_c15FindWithActorIDEjPS_(id, actor);
+                    if (actor == 0)
+                        break;
+                }
+            }
+        }
+        if (*(u8*)(c + 0x602) == 1) {
+            void* a2 = _ZN8dActor_c15FindWithActorIDEjPS_(0xc8, 0);
+            if (a2 != 0) {
+                s32 flag2;
+                if (*(u32*)((char*)a2 + 0xb0) & 0x40000)
+                    flag2 = 1;
+                else
+                    flag2 = 0;
+                if (flag2 != 0) {
+                    *(u8*)AT(c, 0x602) -= 1;
+                }
+            }
+        }
+        if (*(u8*)(c + 0x602) != 0)
+            return;
+        func_02012694(0x15c, c + 0x74);
+        *(u8*)AT(c, 0x603) += 1;
+        *(u8*)(c + 0x602) = 0;
+        return;
+    case 2:
+        if (*(u8*)(c + 0x60b) == 0 && func_ov074_0211f38c((u8*)c) != 0) {
+            *(u8*)(c + 0x60b) = 1;
+        }
+        {
+            s32 r4 = func_ov074_02120474(c);
+            if (func_ov074_021204c0(c) == 0)
+                return;
+            if (r4 == 0)
+                return;
+        }
+        *(u8*)(c + 0x602) = 0;
+        *(u8*)AT(c, 0x603) += 1;
+        return;
+    case 3:
+        if (_ZN6Player18HasFinishedTalkingEv(player) == 0)
+            return;
+        *(u8*)(c + 0x60a) = 0;
+        vzero.x = *(s32*)(c + 0x5c);
+        vzero.y = *(s32*)(c + 0x60);
+        vzero.z = *(s32*)(c + 0x64);
+        vzero.y = vzero.y + 0x190000;
+        _ZN8Particle6System9NewSimpleEj5Fix12IiES2_S2_(0x91, vzero.x, vzero.y, vzero.z);
+        _ZN8Particle6System9NewSimpleEj5Fix12IiES2_S2_(0x92, vzero.x, vzero.y, vzero.z);
+        func_02012694(0x15f, c + 0x74);
+        func_ov074_0211f344(c, 1);
+        func_ov074_0211f244((dActor_c*)c);
+        *(s32*)AT(data0209f318, 0x154) &= ~8;
+        _ZN7Message7EndTalkEv();
+        _ZN5Sound22StopLoadedMusic_Layer3Ev();
+        func_02011cfc();
+        _ZN5Sound17ChangeMusicVolumeEj5Fix12IiE(0x7f, 0x15666);
+        *(u8*)(c + 0x608) = 1;
+        *(u16*)(c + 0x500 + 0xfc) = 0xf;
+        *(u8*)AT(c, 0x603) += 1;
+        return;
+    case 4:
+        vzero.x = *(s32*)(c + 0x5c);
+        vzero.y = *(s32*)(c + 0x60);
+        vzero.z = *(s32*)(c + 0x64);
+        vzero.y = vzero.y + 0x12c000;
+        _ZN8dActor_c5SpawnEjjRK7Vector3PK10Vector3_16as(0x11a, 2, &vzero, 0, *(signed char*)(c + 0xcc), -1);
+        func_02012694(0xbb, c + 0x74);
+        _ZN7fBase_c18MarkForDestructionEv(c);
+        return;
+    default:
+        return;
+    }
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 5 -- func_ov074_0211f38c, 0x0211f38c, size 0x22c */
+/* -------------------------------------------------------------------------- */
+// @symbol func_ov074_0211f38c
+/* recovered: shared common types, declarations from a shared header */
+/* recovered: shared common types */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+int func_ov074_0211f38c(u8* c){
+    struct Vector3 p0, p1;
+    struct Vector3_16 ang;
+    int mult;
+    int r5;
+    void* pl;
+    p0.x=0; p0.y=0; p0.z=0;
+    ang.x=0; ang.y=0; ang.z=0;
+    p0.x=*(int*)(c+0x5c);
+    p0.y=*(int*)(c+0x60);
+    p0.z=*(int*)(c+0x64);
+    p0.y+=0xe4000;
+    p1.x=p0.x;
+    p1.y=p0.y;
+    p1.z=p0.z;
+    ang.y=(short)*(short*)(c+0x8e);
+    if(DecIfAbove0_Short((u16*)(c+0x5fc))==0){
+        u8 t=c[0x602];
+        if(t<=6){
+            ang.y=(short)(t*0x2492+ang.y);
+            mult=0x18c000;
+        } else if(t<=0xb){
+            p0.y+=0xc4000;
+            ang.y=(short)((c[0x602]-7)*0x3333+ang.y);
+            mult=0x108000;
+        } else if(t>0xe){
+            p0.y+=0x24c000;
+            mult=0;
+        } else {
+            p0.y+=0x188000;
+            ang.y=(short)((c[0x602]-0xc)*0x5555+ang.y);
+            mult=0x84000;
+        }
+        {
+            int s=((u16)ang.y>>4)<<1;
+            p0.x += (int)(((s64)mult * data_02082214[s] + 0x800) >> 0xc);
+            p0.z += (int)(((s64)mult * data_02082214[s+1] + 0x800) >> 0xc);
+        }
+        ang.x=Vec3_VertAngle(&p1, &p0);
+        _ZN8dActor_c5SpawnEjjRK7Vector3PK10Vector3_16as(0xc7, 0x1111, &p0, &ang, (signed char)c[0xcc], -1);
+        {
+            u8* pc=(u8*)(c+0x602);
+            *pc+=1;
+        }
+        *(short*)((c+0x500)+0xfc)=2;
+    }
+    r5=data_ov074_02122d80[c[0x604]];
+    pl=_ZN8dActor_c13ClosestPlayerEv(c);
+    if(c[0x604]==1){ if(*(int*)((char*)pl+8)!=3) r5+=1; }
+    return c[0x602]>=r5 ? 1 : 0;
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 4 -- func_ov074_0211f344, 0x0211f344, size 0x48 */
+/* -------------------------------------------------------------------------- */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+void func_ov074_0211f344(void* r6, unsigned char r5) {
+  void* r1 = (void*)_ZN8dActor_c15FindWithActorIDEjPS_(0xc7, 0);
+  while (r1) {
+    if (r1 != r6) *(unsigned char*)((char*)r1 + 0x60a) = r5;
+    r1 = (void*)_ZN8dActor_c15FindWithActorIDEjPS_(0xc7, r1);
+  }
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 3 -- func_ov074_0211f244, 0x0211f244, size 0x100 */
+/* -------------------------------------------------------------------------- */
+// @symbol func_ov074_0211f244
+/* recovered: shared common types */
+extern "C" void func_ov074_0211f244(dActor_c *self)
+{
+    dActor_c *a = dActor_c::FindWithActorID(0xc7, 0);
+    if (a == 0) return;
+    do {
+        if (a != self) {
+            char *t = (char*)a;
+            *(int*)(t + 0x9c) = -0xa000;
+            int rv = (unsigned int)RandomIntInternal(&data_0209e650) >> 8;
+            *(short*)(t + 0x5fc) = 0x30;
+            *(int*)(t + 0x98) = ((rv & 0xf) << 0xc) + 0x16000;
+            *(int*)(t + 0xa4) = 0;
+            *(int*)(t + 0xa8) = ((rv & 0x1f) << 0xc) + 0x60000;
+            *(int*)(t + 0xac) = 0;
+            Vector3_16 pos = *(Vector3_16*)(t + 0x92);
+            rv &= 0x1000;
+            pos.y = (short)(pos.y + rv);
+            short x = pos.x + 0;
+            *(short*)(t + 0x92) = x;
+            short y = pos.y + 0;
+            *(short*)(t + 0x94) = y;
+            short z = pos.z + 0;
+            *(short*)(t + 0x96) = z;
+            *(short*)(t + 0x8c) = x;
+            *(short*)(t + 0x8e) = y;
+            *(short*)(t + 0x90) = z;
+            *(unsigned char*)(t + 0x609) = 1;
+        }
+        a = dActor_c::FindWithActorID(0xc7, a);
+    } while (a != 0);
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 2 -- func_ov074_0211f154, 0x0211f154, size 0xf0 */
+/* -------------------------------------------------------------------------- */
+// @symbol func_ov074_0211f154
+/* recovered: shared common types */
+/* func_ov074_0211f154 at 0x0211f154
+ *
+ * Matched byte-for-byte with mwccarm 1.2/sp2p3 (ov074).
+ */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+void func_ov074_0211f154(char* c) {
+    struct Vector3 look, pos, in, out;
+    void* cam;
+    cam = data_0209f318;
+    _ZN6Camera9SetFlag_3Ev(cam);
+    in.x = 0; in.y = 0; in.z = 0;
+    out.x = 0; out.y = 0; out.z = 0;
+    look.x = *(int*)(c + 0x5c);
+    look.y = *(int*)(c + 0x60);
+    look.z = *(int*)(c + 0x64);
+    pos.x = *(int*)(c + 0x5c);
+    pos.y = *(int*)(c + 0x60);
+    pos.z = *(int*)(c + 0x64);
+    look.y += 0x1a0000;
+    in.z = 0x59c000;
+    Matrix4x3_FromRotationY(&data_020a0e68, *(short*)(c + 0x8e));
+    MulVec3Mat4x3(&in, &data_020a0e68, &out);
+    pos.x = pos.x + out.x;
+    pos.y = pos.y + 0x300000;
+    pos.z = pos.z + out.z;
+    _ZN6Camera9SetLookAtERK7Vector3(cam, &look);
+    _ZN6Camera6SetPosERK7Vector3(cam, &pos);
+}
+}
+
