--- src/func_ov074_0211fc34.c
@@ -1,3 +1,0 @@
-void func_ov074_0211fc34(void)
-{
-}
