--- src/func_ov074_0212007c.c
@@ -1,3 +1,0 @@
-void func_ov074_0212007c(void)
-{
-}
