--- src/func_ov074_0212016c.c
@@ -1,7 +1,0 @@
-extern void func_ov074_02121a4c(char* c, int idx);
-extern void func_ov074_02120b90(void *c);
-
-void func_ov074_0212016c(void *c) {
-    func_ov074_02121a4c((char*)c, 0xa);
-    func_ov074_02120b90(c);
-}
