--- src/func_ov074_021201ec.c
@@ -1,3 +1,0 @@
-void func_ov074_021201ec(void)
-{
-}
