--- src/func_ov074_021203e0.c
@@ -1,3 +1,0 @@
-void func_ov074_021203e0(void)
-{
-}
