--- src/func_ov074_02120808.cpp
@@ -1,15 +1,0 @@
-//cpp
-extern "C" {
-int _ZN8dActor_c18HorzAngleToCPlayerEv(void* p);
-int AngleDiff(int a, int b);
-void func_ov074_02121a4c(void* c, int i);
-}
-
-extern "C" void func_ov074_02120808(char* c)
-{
-    int ang = (short)(*(short*)(c + 0x5f4) + *(int*)(c + 0x5f0) * 0x4500);
-    if (AngleDiff(ang, _ZN8dActor_c18HorzAngleToCPlayerEv(c)) <= 0x4000)
-        return;
-    func_ov074_02121a4c(c, 6);
-    *(int *)(((int)(c + 0x5f0))) *= -1;
-}
